requirements.txt
cohere==4.37
gitpython==3.1.42
python-dotenv==1.0.1
pathlib 